requirements.txt
Flask
google-genai
numpy
pandas
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-bigquery
pandas-gbq
